Split cartridge code into modules
